Added structures to new file
